+++ packages/hub/src/scope.ts
@@ -2,6 +2,8 @@
 import {
   Breadcrumb,
   CaptureContext,
+  Context,
+  Contexts,
   Event,
   EventHint,
   EventProcessor,
@@ -14,7 +16,7 @@ import {
   Transaction,
   User,
 } from '@sentry/types';
-import { getGlobalObject, isPlainObject, isThenable, SyncPromise, timestampWithMs } from '@sentry/utils';
+import { dateTimestampInSeconds, getGlobalObject, isPlainObject, isThenable, SyncPromise } from '@sentry/utils';
 
 /**
  * Holds additional event information. {@link Scope.applyToEvent} will be
@@ -40,12 +42,10 @@ export class Scope implements ScopeInterface {
   protected _tags: { [key: string]: string } = {};
 
   /** Extra */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  protected _extra: { [key: string]: any } = {};
+  protected _extra: Extras = {};
 
   /** Contexts */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  protected _contexts: { [key: string]: any } = {};
+  protected _contexts: Contexts = {};
 
   /** Fingerprint */
   protected _fingerprint?: string[];
@@ -185,9 +185,14 @@ export class Scope implements ScopeInterface {
   /**
    * @inheritDoc
    */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  public setContext(key: string, context: { [key: string]: any } | null): this {
-    this._contexts = { ...this._contexts, [key]: context };
+  public setContext(key: string, context: Context | null): this {
+    if (context === null) {
+      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+      delete this._contexts[key];
+    } else {
+      this._contexts = { ...this._contexts, [key]: context };
+    }
+
     this._notifyScopeListeners();
     return this;
   }
@@ -212,10 +217,20 @@ export class Scope implements ScopeInterface {
    * @inheritDoc
    */
   public getTransaction(): Transaction | undefined {
-    const span = this.getSpan() as Span & { spanRecorder: { spans: Span[] } };
-    if (span && span.spanRecorder && span.spanRecorder.spans[0]) {
+    // often, this span will be a transaction, but it's not guaranteed to be
+    const span = this.getSpan() as undefined | (Span & { spanRecorder: { spans: Span[] } });
+
+    // try it the new way first
+    if (span?.transaction) {
+      return span?.transaction;
+    }
+
+    // fallback to the old way (known bug: this only finds transactions with sampled = true)
+    if (span?.spanRecorder?.spans[0]) {
       return span.spanRecorder.spans[0] as Transaction;
     }
+
+    // neither way found a transaction
     return undefined;
   }
 
@@ -287,7 +302,7 @@ export class Scope implements ScopeInterface {
    */
   public addBreadcrumb(breadcrumb: Breadcrumb, maxBreadcrumbs?: number): this {
     const mergedBreadcrumb = {
-      timestamp: timestampWithMs(),
+      timestamp: dateTimestampInSeconds(),
       ...breadcrumb,
     };